Split RemoteQueryExecutorReadContext into module part
