+++ src/Common/FiberStack.h
@@ -3,6 +3,7 @@
 #include <boost/context/stack_context.hpp>
 #include <Common/formatReadable.h>
 #include <Common/CurrentMemoryTracker.h>
+#include <Common/Exception.h>
 
 #include <sys/time.h>
 #include <sys/resource.h>
